Purchase OrderImplementation updates

+++ src/main/java/com/sams/inventorymanagement/services/StoreServiceImpl.java
@@ -1,8 +1,10 @@
 package com.sams.inventorymanagement.services;
 
+import com.sams.inventorymanagement.entities.Inventory;
 import com.sams.inventorymanagement.entities.Store;
 import com.sams.inventorymanagement.enums.StoreType;
 import com.sams.inventorymanagement.exceptions.EntityDuplicateException;
+import com.sams.inventorymanagement.repositories.InventoryRepository;
 import com.sams.inventorymanagement.repositories.StoreRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -14,6 +16,9 @@ import java.util.List;
 public class StoreServiceImpl implements StoreService {
     @Autowired
     private StoreRepository storeRepository;
+
+    @Autowired
+    private InventoryRepository inventoryRepository;
     
     @Override
     public List<Store> searchStoresByCriteria(String name, String address, StoreType storeType, LocalDate openingDate) {
@@ -41,14 +46,30 @@ public class StoreServiceImpl implements StoreService {
 
     @Override
     public void deleteStore(Long id) {
+        Store store = storeRepository.findById(id).orElse(null);
+
+        if (store == null) return;
+
+//        TODO WIP
+        List<Inventory> inventories = store.getInventories();
+        for (Inventory inventory : inventories) {
+            inventory.setStore(null);
+            inventoryRepository.delete(inventory);
+        }
+
         storeRepository.deleteById(id);
     }
 
     @Override
     public Store updateStore(Long id, Store updatedStore) {
-        if (storeRepository.existsById(id)) {
-            updatedStore.setId(id); // Ensure the ID is set
-            return storeRepository.save(updatedStore);
+        Store store = storeRepository.findById(id).orElse(null);
+        if (store != null) {
+            store.setName(updatedStore.getName());
+            store.setAddress(updatedStore.getAddress());
+            store.setEmail(updatedStore.getEmail());
+            store.setPhone(updatedStore.getPhone());
+            store.setType(updatedStore.getType());
+            return storeRepository.save(store);
         } else {
             return null; // Return null if the store with the specified ID does not exist
         }
